+++ api/web3server_utils.go
@@ -19,6 +19,7 @@ import (
 	"github.com/iotexproject/go-pkgs/util"
 	"github.com/iotexproject/iotex-address/address"
 	"github.com/iotexproject/iotex-proto/golang/iotexapi"
+	"github.com/iotexproject/iotex-proto/golang/iotextypes"
 	"github.com/pkg/errors"
 	"github.com/tidwall/gjson"
 	"go.uber.org/zap"
@@ -468,4 +469,42 @@ func fromLoggerStructLogs(logs []logger.StructLog) []apitypes.StructLog {
 		}
 	}
 	return ret
+}
+
+func newGetTransactionResult(
+	blkHash hash.Hash256,
+	selp action.SealedEnvelope,
+	receipt *action.Receipt,
+	evmChainID uint32,
+) (*getTransactionResult, error) {
+	act, ok := selp.Action().(action.EthCompatibleAction)
+	if !ok {
+		actHash, _ := selp.Hash()
+		return nil, errors.Wrapf(errUnsupportedAction, "actHash: %s", hex.EncodeToString(actHash[:]))
+	}
+	ethTx, err := act.ToEthTx(0)
+	if err != nil {
+		return nil, err
+	}
+	var to *string
+	if ethTx.To() != nil {
+		tmp := ethTx.To().String()
+		to = &tmp
+	}
+
+	signer, err := action.NewEthSigner(iotextypes.Encoding(selp.Encoding()), evmChainID)
+	if err != nil {
+		return nil, err
+	}
+	tx, err := action.RawTxToSignedTx(ethTx, signer, selp.Signature())
+	if err != nil {
+		return nil, err
+	}
+	return &getTransactionResult{
+		blockHash: blkHash,
+		to:        to,
+		ethTx:     tx,
+		receipt:   receipt,
+		pubkey:    selp.SrcPubkey(),
+	}, nil
 }
